bot/requirements.txt
aiogram==3.13.1
fastapi==0.115.6
uvicorn==0.34.0
asyncpg==0.30.0
httpx==0.28.1
pydantic==2.9.2
redis==5.2.1
python-dotenv==1.0.1
